finish transfet to nodejs
