+++ client/src/app/auth/signin/page.tsx
@@ -1,9 +1,16 @@
 import { getProviders } from "next-auth/react";
-import SignInClient from "@/components/signin/SignInClient";
+import SigninContent from "@/components/auth/SigninContent";
 
 export default async function SignInPage() {
   const providers = await getProviders();
-  if (providers == null) return;
 
-  return <SignInClient providers={providers} />;
+  if (!providers) {
+    return (
+      <div className="w-full h-full flex items-center justify-center">
+        <span>No providers provided.</span>
+      </div>
+    );
+  }
+
+  return <SigninContent providers={providers} />;
 }
