+++ client/src/context/ContextProviderWrapper.tsx
@@ -3,8 +3,12 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { SessionProvider } from "next-auth/react";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+import axios from "axios";
 
 const queryClient = new QueryClient();
+
+axios.defaults.withCredentials = true;
+axios.defaults.baseURL = process.env.NEXT_PUBLIC_API_URL;
 
 export default function QueryClientProviderWrapper({
   children,
