--- client/src/models/offer.ts
@@ -1,18 +1,0 @@
-import { Schema, model, models } from "mongoose";
-
-const OfferSchema = new Schema(
-  {
-    name: { type: String, required: true },
-    url: { type: String, required: true },
-    lastPrices: { type: [String], required: false, default: [] },
-    currentPrice: { type: String, required: false, default: "" },
-    userId: { type: String, required: true },
-    status: { type: String, required: false, default: "notChanged" },
-    img: { type: String, required: false, default: "" },
-  },
-  { timestamps: true }
-);
-
-OfferSchema.index({ userId: 1, url: 1 }, { unique: true });
-
-export const Offer = models.Offer || model("Offer", OfferSchema);
